+++ ui/molecules/InputCheck.tsx
@@ -1,22 +1,20 @@
-//   ___                _   ___ _     _    _  ___           _
-//  |_ _|_ _  _ __ _  _| |_| __(_)___| |__| |/ __|___ _ __ | |__  ___
-//   | || ' \| '_ \ || |  _| _|| / -_) / _` | (__/ _ \ '  \| '_ \/ _ \
-//  |___|_||_| .__/\_,_|\__|_| |_\___|_\__,_|\___\___/_|_|_|_.__/\___/
+//   ___                _    ___ _           _    ___           _
+//  |_ _|_ _  _ __ _  _| |_ / __| |_  ___ __| |__/ __|___ _ __ | |__  ___
+//   | || ' \| '_ \ || |  _| (__| ' \/ -_) _| / / (__/ _ \ '  \| '_ \/ _ \
+//  |___|_||_| .__/\_,_|\__|\___|_||_\___\__|_\_\\___\___/_|_|_|_.__/\___/
 //           |_|
 ////////////////////////////////////////////////////////////////////////////////////////////////////////
 /**
- * Module is for the `InputFieldCombo` molecule.
- * @module molecules/InputFieldCombo
+ * Module is for the `InputCheck` molecule.
+ * @module molecules/InputCheck
  */
 import type { JSX } from 'preact';
-import { cn } from '@vyn/cn';
 import { apDef } from '@lunchbox/ui';
-import { input } from '../particles.ts';
 import Input from '../atoms/Input.tsx';
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-/** Property interface for the `InputFieldCombo` molecule. */
-export interface iInputFieldCombo {
+/** Property interface for the `InputCheck` molecule. */
+export interface iInputCheck {
   /**
    * The input's title. It is built using the `<label/>` element already without the need of linking
    * them with the ids.
@@ -30,15 +28,22 @@ export interface iInputFieldCombo {
   error: string;
 
   /**
+   * The identifier that must be shared by other checkbox inputs inside so that they're related in the
+   * form's context.
+   */
+  name: string;
+
+  /**
    * Makes this input required during form submission.
    */
   required: boolean;
 }
 
-/** Default properties of the `InputFieldCombo` molecule. */
-const d: iInputFieldCombo = {
+/** Default properties of the `InputCheck` molecule. */
+const d: iInputCheck = {
   label: '',
   error: '',
+  name: '',
   required: false,
 };
 
@@ -48,34 +53,24 @@ const d: iInputFieldCombo = {
  *
  * @todo [DOC] Add a description and code example of how to propertly use this component.
  * @todo [DOC] Explain the relationship this molecule has with the Input atom.
- * @todo [DEV] Handle dates an times in useful notations.
  *
  * @example
  * ```ts
- * import { InputFieldCombo } from 'lunchbox/molecules';
+ * import { InputCheck } from 'lunchbox/molecules';
  *
  * (example code)
  * ```
  */
-export default function (props: Partial<iInputFieldCombo>): JSX.Element {
+export default function (props: Partial<iInputCheck>): JSX.Element {
   const p = apDef(d, props);
 
   return (
-    <Input.Container>
-      <Input.Label class='flex-col'>
-        {p.label
-          ? (
-            <Input.Text class={p.required ? input.required : undefined}>
-              {p.label}
-            </Input.Text>
-          )
-          : null}
-        <Input.Field
-          required={p.required}
-          class={cn(p.error ? input.invalid : null)}
-        />
+    <div>
+      <Input.Label class='items-center'>
+        <Input.Check name={p.name} required={p.required} />
+        <Input.Text>{p.label}</Input.Text>
       </Input.Label>
       {p.error ? <Input.Error>{p.error}</Input.Error> : null}
-    </Input.Container>
+    </div>
   );
 }